+++ lib/src/theme/sbb_typography.dart
@@ -11,10 +11,12 @@ class SBBTextStyles {
   static const double largeFontSize = 18.0;
   static const double mediumFontSize = 16.0;
   static const double smallFontSize = 14.0;
+  static const double xSmallFontSize = 12.0;
 
   static const double largeFontHeight = 24.0 / largeFontSize;
   static const double mediumFontHeight = 20.0 / mediumFontSize;
   static const double smallFontHeight = 20.0 / smallFontSize;
+  static const double xSmallFontHeight = 16.0 / smallFontSize;
 
   static const TextStyle extraLargeLight = TextStyle(
     fontSize: 30.0,
@@ -73,10 +75,18 @@ class SBBTextStyles {
   );
 
   static const TextStyle extraSmallLight = TextStyle(
-    fontSize: 12.0,
-    height: 16.0 / 12.0,
+    fontSize: xSmallFontSize,
+    height: xSmallFontHeight,
     fontStyle: FontStyle.normal,
     fontWeight: FontWeight.w300,
+    fontFamily: SBBWebFont,
+  );
+
+  static const TextStyle extraSmallBold = TextStyle(
+    fontSize: xSmallFontSize,
+    height: xSmallFontHeight,
+    fontStyle: FontStyle.normal,
+    fontWeight: FontWeight.w700,
     fontFamily: SBBWebFont,
   );